adjust project
