+++ truco_mobile/lib/main.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart' hide Table, Card;
 import 'package:truco_mobile/src/config/general_config.dart';
-import 'package:truco_mobile/src/model/Card/cardmodel.dart';
+import 'package:truco_mobile/src/model/cardmodel.dart';
 import 'package:truco_mobile/src/service/api_service.dart';
 //import 'package:truco_mobile/src/model/Player/player_model.dart';
 //import 'package:truco_mobile/src/model/Table/table_model.dart';
-import 'package:truco_mobile/src/view/BoardView/board_view.dart';
-import 'package:truco_mobile/src/view/LoadingView/loading_view.dart';
-import 'package:truco_mobile/src/view/LoginView/login_view.dart';
-import 'package:truco_mobile/src/view/MainView/main_page_view.dart';
-import 'package:truco_mobile/src/view/RegisterView/register_view.dart';
+import 'package:truco_mobile/src/view/board_view.dart';
+import 'package:truco_mobile/src/view/loading_view.dart';
+import 'package:truco_mobile/src/view/login_view.dart';
+import 'package:truco_mobile/src/view/main_page_view.dart';
+import 'package:truco_mobile/src/view/register_view.dart';
 //import 'package:truco_mobile/src/view/game_screen.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
