+++ truco_mobile/lib/src/config/general_config.dart
@@ -1,2 +1,44 @@
 const BACK_CARD = 'https://www.deckofcardsapi.com/static/img/back.png';
 const DECK_API = 'https://deckofcardsapi.com/api';
+const DECK_API_CARDS = [
+  'AD',
+  'AS',
+  'AH',
+  'AC',
+  '2D',
+  '2S',
+  '2H',
+  '2C',
+  '3D',
+  '3S',
+  '3H',
+  '3C',
+  '4D',
+  '4S',
+  '4H',
+  '4C',
+  '5D',
+  '5S',
+  '5H',
+  '5C',
+  '6D',
+  '6S',
+  '6H',
+  '6C',
+  '7D',
+  '7S',
+  '7H',
+  '7C',
+  'QD',
+  'QS',
+  'QH',
+  'QC',
+  'JD',
+  'JS',
+  'JH',
+  'JC',
+  'KD',
+  'KS',
+  'KH',
+  'KC'
+];
