+++ truco_mobile/lib/src/model/card_model.dart
