+++ truco_mobile/lib/src/model/cardmodel.dart
