+++ truco_mobile/lib/src/view/game_list_view.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:truco_mobile/src/config/general_config.dart';
-import 'package:truco_mobile/src/controller/truco_controller.dart';
 import 'package:truco_mobile/src/service/api_service.dart';
-import 'package:truco_mobile/src/view/BoardView/board_view.dart';
-import 'package:truco_mobile/src/view/HomeView/home_view.dart';
-import 'package:truco_mobile/src/model/Card/cardmodel.dart';
+import 'package:truco_mobile/src/view/board_view.dart';
+import 'package:truco_mobile/src/view/home_view.dart';
+import 'package:truco_mobile/src/model/cardmodel.dart';
 
 class GameListView extends StatefulWidget {
   final String title;
@@ -16,8 +15,7 @@ class GameListView extends StatefulWidget {
 }
 
 class _GameListView extends State<GameListView> {
-  TrucoController trucoController =
-      TrucoController(apiService: ApiService(baseUrl: DECK_API));
+  // TrucoController trucoController = TrucoController(apiService: ApiService(baseUrl: DECK_API));
 
   List<Map<String, dynamic>> gameRooms = [
     {'name': 'Sala 1', 'players': 4, 'maxPlayers': 4},
@@ -113,13 +111,13 @@ class _GameListView extends State<GameListView> {
         title: Text(name),
         trailing: Text('$players/$maxPlayers jogadores'),
         onTap: () async {
-          if (name == 'Sala 1') {
-            List<CardModel> cards = await trucoController.fetchCards();
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => BoardView(cards: cards)),
-            );
-          }
+          // if (name == 'Sala 1') {
+          //   List<CardModel> cards = await trucoController.fetchCards();
+          //   Navigator.push(
+          //     context,
+          //     MaterialPageRoute(builder: (context) => BoardView(cards: cards)),
+          //   );
+          // }
         },
       ),
     );
