+++ truco_mobile/lib/src/view/loading_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:truco_mobile/src/view/MainView/main_page_view.dart';
+import 'package:truco_mobile/src/view/main_page_view.dart';
 
 class MyLoadingPage extends StatefulWidget {
   final String title;
