+++ truco_mobile/lib/src/view/login_view.dart
@@ -1,41 +1,46 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:truco_mobile/src/view/MainView/main_page_view.dart';
-import 'package:truco_mobile/src/widget/CustomButton/custom_button.dart';
-import 'package:truco_mobile/src/widget/CustomTextInput/custom_text.dart';
+import 'package:truco_mobile/src/view/home_view.dart';
+import 'package:truco_mobile/src/view/main_page_view.dart';
+import 'package:truco_mobile/src/widget/custom_button.dart';
+import 'package:truco_mobile/src/widget/custom_text.dart';
 
-class MyRegisterPage extends StatefulWidget {
+class MyLoginPage extends StatefulWidget {
   final String title;
 
-  const MyRegisterPage({Key? key, required this.title}) : super(key: key);
+  const MyLoginPage({Key? key, required this.title}) : super(key: key);
 
   @override
-  _MyRegisterPageState createState() => _MyRegisterPageState();
+  _MyLoginPageState createState() => _MyLoginPageState();
 }
 
-class _MyRegisterPageState extends State<MyRegisterPage> {
-  final TextEditingController nameController = TextEditingController();
+class _MyLoginPageState extends State<MyLoginPage> {
   final TextEditingController emailController = TextEditingController();
   final TextEditingController passwordController = TextEditingController();
   final FirebaseAuth _auth = FirebaseAuth.instance;
-
-  Future<void> _register() async {
+  bool isDevelopment = true;
+  Future<void> _login() async {
+    
+    if (isDevelopment) {
+      Navigator.push(context, MaterialPageRoute(builder: (context) => MyHomePagePage(title: 'Tela inicial')));
+    }
+    
     try {
-      UserCredential userCredential = await _auth.createUserWithEmailAndPassword(
+      UserCredential userCredential = await _auth.signInWithEmailAndPassword(
         email: emailController.text,
         password: passwordController.text,
       );
       User? user = userCredential.user;
       if (user != null) {
-        // Navegar para a tela principal após o registro bem-sucedido
+        // Navegar para a tela inicial após o login bem-sucedido
         Navigator.pushReplacement(
           context,
-          MaterialPageRoute(builder: (context) => MyMainPagePage()),
+          MaterialPageRoute(builder: (context) => MyHomePagePage(title: 'Tela inicial')),
         );
       }
     } on FirebaseAuthException catch (e) {
-      // Mostrar um erro se o registro falhar
-      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.message ?? 'Erro ao registrar')));
+      // Mostrar um erro se o login falhar
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.message ?? 'Erro ao fazer login')));
     } catch (e) {
       print(e);
     }
@@ -71,14 +76,6 @@ class _MyRegisterPageState extends State<MyRegisterPage> {
                 child: Scrollbar(
                   child: ListView(
                     children: <Widget>[
-                      CustomTextField(
-                        controller: nameController,
-                        hintText: "Digite seu nome...",
-                        labelTextFontSize: 36,
-                        hintTextFontSize: 20,
-                        fontSize: 20,
-                        labelText: "Nome",
-                      ),
                       SizedBox(height: 30),
                       CustomTextField(
                         controller: emailController,
@@ -100,8 +97,8 @@ class _MyRegisterPageState extends State<MyRegisterPage> {
                       ),
                       SizedBox(height: 30),
                       CustomButton(
-                        text: 'Salvar',
-                        onPressed: _register,
+                        text: 'Entrar',
+                        onPressed: _login,
                         width: 50,
                         height: 50,
                         fontSize: 36,
