+++ truco_mobile/lib/src/view/register_view.dart
@@ -1,41 +1,41 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:truco_mobile/src/view/HomeView/home_view.dart';
-import 'package:truco_mobile/src/view/MainView/main_page_view.dart';
-import 'package:truco_mobile/src/widget/CustomButton/custom_button.dart';
-import 'package:truco_mobile/src/widget/CustomTextInput/custom_text.dart';
+import 'package:truco_mobile/src/view/main_page_view.dart';
+import 'package:truco_mobile/src/widget/custom_button.dart';
+import 'package:truco_mobile/src/widget/custom_text.dart';
 
-class MyLoginPage extends StatefulWidget {
+class MyRegisterPage extends StatefulWidget {
   final String title;
 
-  const MyLoginPage({Key? key, required this.title}) : super(key: key);
+  const MyRegisterPage({Key? key, required this.title}) : super(key: key);
 
   @override
-  _MyLoginPageState createState() => _MyLoginPageState();
+  _MyRegisterPageState createState() => _MyRegisterPageState();
 }
 
-class _MyLoginPageState extends State<MyLoginPage> {
+class _MyRegisterPageState extends State<MyRegisterPage> {
+  final TextEditingController nameController = TextEditingController();
   final TextEditingController emailController = TextEditingController();
   final TextEditingController passwordController = TextEditingController();
   final FirebaseAuth _auth = FirebaseAuth.instance;
 
-  Future<void> _login() async {
+  Future<void> _register() async {
     try {
-      UserCredential userCredential = await _auth.signInWithEmailAndPassword(
+      UserCredential userCredential = await _auth.createUserWithEmailAndPassword(
         email: emailController.text,
         password: passwordController.text,
       );
       User? user = userCredential.user;
       if (user != null) {
-        // Navegar para a tela inicial após o login bem-sucedido
+        // Navegar para a tela principal após o registro bem-sucedido
         Navigator.pushReplacement(
           context,
-          MaterialPageRoute(builder: (context) => MyHomePagePage(title: 'Tela inicial')),
+          MaterialPageRoute(builder: (context) => MyMainPagePage()),
         );
       }
     } on FirebaseAuthException catch (e) {
-      // Mostrar um erro se o login falhar
-      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.message ?? 'Erro ao fazer login')));
+      // Mostrar um erro se o registro falhar
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(e.message ?? 'Erro ao registrar')));
     } catch (e) {
       print(e);
     }
@@ -71,6 +71,14 @@ class _MyLoginPageState extends State<MyLoginPage> {
                 child: Scrollbar(
                   child: ListView(
                     children: <Widget>[
+                      CustomTextField(
+                        controller: nameController,
+                        hintText: "Digite seu nome...",
+                        labelTextFontSize: 36,
+                        hintTextFontSize: 20,
+                        fontSize: 20,
+                        labelText: "Nome",
+                      ),
                       SizedBox(height: 30),
                       CustomTextField(
                         controller: emailController,
@@ -92,8 +100,8 @@ class _MyLoginPageState extends State<MyLoginPage> {
                       ),
                       SizedBox(height: 30),
                       CustomButton(
-                        text: 'Entrar',
-                        onPressed: _login,
+                        text: 'Salvar',
+                        onPressed: _register,
                         width: 50,
                         height: 50,
                         fontSize: 36,
