+++ truco_mobile/lib/src/widget/custom_text.dart
