+++ truco_mobile/lib/src/widget/score_board.dart
