+++ truco_mobile/test/team_model_test.dart
@@ -1,33 +1,33 @@
-// ignore_for_file: unused_local_variable
+// // ignore_for_file: unused_local_variable
 
-import 'package:test/test.dart';
-import 'package:truco_mobile/src/model/Team/team_model.dart';
-import 'package:truco_mobile/src/model/Player/player_model.dart';
-import 'package:truco_mobile/src/model/Deck/deck_model.dart';
+// import 'package:test/test.dart';
+// import 'package:truco_mobile/src/model/team_model.dart';
+// import 'package:truco_mobile/src/model/player_model.dart';
+// import 'package:truco_mobile/src/model/deck_model.dart';
 
-void main() {
-  late Team team;
-  late Deck deck;
+// void main() {
+//   late Team team;
+//   late Deck deck;
 
-  group(
-      'Team',
-      () => {
-            setUp(() => {
-                  team = Team(
-                      name: 'Team 1', players: [Player(name: 'Player 1'),Player(name: 'Player 2')]),
-                  deck = Deck()
-                }),
-            test('should initialize with 0 points', () {
-              expect(team.getPoints(), 0);
-            }),
-            test('should add points', () {
-              team.addPoints(1);
-              expect(team.getPoints(), 1);
-            }),
-            test('should reset points', () {
-              team.addPoints(12);
-              team.resetPoints();
-              expect(team.getPoints(), 0);
-            }),
-          });
-}
+//   group(
+//       'Team',
+//       () => {
+//             setUp(() => {
+//                   team = Team(
+//                       name: 'Team 1', players: [Player(name: 'Player 1'),Player(name: 'Player 2')]),
+//                   deck = Deck()
+//                 }),
+//             test('should initialize with 0 points', () {
+//               expect(team.getPoints(), 0);
+//             }),
+//             test('should add points', () {
+//               team.addPoints(1);
+//               expect(team.getPoints(), 1);
+//             }),
+//             test('should reset points', () {
+//               team.addPoints(12);
+//               team.resetPoints();
+//               expect(team.getPoints(), 0);
+//             }),
+//           });
+// }
