+++ truco_mobile/test/trucoStatus_model_test.dart
@@ -1,16 +1,16 @@
-import 'package:test/test.dart';
-import 'package:truco_mobile/src/model/trucoStatus_model.dart'; // Substitua "seu_pacote" pelo nome do seu pacote
+// import 'package:test/test.dart';
+// import 'package:truco_mobile/src/model/trucoStatus_model.dart'; // Substitua "seu_pacote" pelo nome do seu pacote
 
-void main() {
-  group('TrucoStatusExtension', () {
-    test('pointsAwarded() should return correct points for each status', () {
-      expect(TrucoStatus.NOT_REQUESTED.pointsAwarded(), equals(0));
-      expect(TrucoStatus.REQUESTED_3_POINTS.pointsAwarded(), equals(3));
-      expect(TrucoStatus.REQUESTED_6_POINTS.pointsAwarded(), equals(6));
-      expect(TrucoStatus.REQUESTED_9_POINTS.pointsAwarded(), equals(9));
-      expect(TrucoStatus.REQUESTED_12_POINTS.pointsAwarded(), equals(12));
-      expect(TrucoStatus.ACCEPTED.pointsAwarded(), equals(0));
-      expect(TrucoStatus.REFUSED.pointsAwarded(), equals(0));
-    });
-  });
-}
+// void main() {
+//   group('TrucoStatusExtension', () {
+//     test('pointsAwarded() should return correct points for each status', () {
+//       expect(TrucoStatus.NOT_REQUESTED.pointsAwarded(), equals(0));
+//       expect(TrucoStatus.REQUESTED_3_POINTS.pointsAwarded(), equals(3));
+//       expect(TrucoStatus.REQUESTED_6_POINTS.pointsAwarded(), equals(6));
+//       expect(TrucoStatus.REQUESTED_9_POINTS.pointsAwarded(), equals(9));
+//       expect(TrucoStatus.REQUESTED_12_POINTS.pointsAwarded(), equals(12));
+//       expect(TrucoStatus.ACCEPTED.pointsAwarded(), equals(0));
+//       expect(TrucoStatus.REFUSED.pointsAwarded(), equals(0));
+//     });
+//   });
+// }
